added project management services to create project, task, subtask and list project

+++ Models/AppModels/ProjectTimesheet.cs
@@ -10,7 +10,7 @@ namespace TimesheetBE.Models.AppModels
         public Guid? SubTaskId { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public int PercentageOfCompletion { get; set; }
+        public double PercentageOfCompletion { get; set; }
         public bool Billable { get; set; }
     }
 }
